move Persistor to own file
