+++ src/js/reducers/credentials.js
@@ -1,7 +1,7 @@
 // @flow
+import Persistor from '../modules/Persistor'
 import type { Action, CredentialsT } from '../flow'
 type State = CredentialsT[]
-const { localStorage } = window
 
 export function credentialsToId({
   realName,
@@ -9,43 +9,6 @@ export function credentialsToId({
   port
 }: CredentialsT): string {
   return `${realName}@${server}:${port}`
-}
-
-class Persistor<A> {
-  key: string
-  default: A
-
-  constructor(key: string, def: A) {
-    this.key = key
-    this.default = def
-  }
-
-  init(): A {
-    if (this.key in localStorage) {
-      try {
-        return JSON.parse(localStorage[this.key])
-      } catch (e) {
-        if (e instanceof SyntaxError) {
-          localStorage.removeItem(this.key)
-          return this.default
-        } else {
-          throw e
-        }
-      }
-    } else {
-      return this.default
-    }
-  }
-
-  wrap(reducer: (state: A, action: Action) => A) {
-    return (state: A, action: Action) => {
-      const after = reducer(state || this.init(), action)
-      if (typeof after !== 'undefined' && state !== after) {
-        localStorage.setItem(this.key, JSON.stringify(after))
-      }
-      return after
-    }
-  }
 }
 
 const persist: Persistor<State> = new Persistor('past-credentials', ([]: State))
