+++ src/static/connection-view.css
@@ -1,0 +1,70 @@
+.connection-view {
+  margin-left: 60px; }
+
+.login-form {
+  display: flex;
+  flex-direction: column;
+  width: 220px;
+  margin: 0 auto; }
+
+.login-form .checkbox-label {
+  margin: 24px 0 12px;
+  text-align: center; }
+
+.connection-error {
+  text-align: center;
+  padding: 12px 0;
+  margin: 18px 0;
+  color: #721c24;
+  background-color: #f8d7da;
+  border-color: #f5c6cb; }
+
+.disconnected-error {
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+  display: flex;
+  align-items: center;
+  margin: 1rem;
+  padding: 0.2rem 0.6rem;
+  color: #383c4a;
+  background: #f2f2f5;
+  font-weight: bold;
+  font-size: 18px;
+  border-radius: 3px; }
+
+.disconnected-error svg {
+  margin-right: 6px; }
+
+.conversation-list {
+  background: #373d48;
+  color: #fff; }
+
+.connection-creator .right-panel,
+.connection-creator .full-panel {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  height: 100vh; }
+
+.connection-creator .left-panel {
+  border-right: 1px solid #ccc; }
+
+.saved-credentials {
+  display: block;
+  position: relative;
+  width: 100%;
+  text-align: left;
+  padding: 12px 18px;
+  outline: none;
+  cursor: default; }
+
+.saved-credentials:hover .nickname {
+  text-decoration: underline; }
+
+.saved-credentials .forget {
+  position: absolute;
+  right: 18px;
+  top: 9px;
+  font-size: 18px;
+  font-family: monospace; }
